pyproject.toml
[project]
name = "fktz"
version = "0.1.1"
description = "A discord bot for Discord Timestamps – because f*ck timezones."
readme = "README.md"
authors = [
    { name = "Veivel", email = "25278964+Veivel@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "python-dateutil>=2.9.0.post0",
    "ruff>=0.12.7",
]

[project.scripts]
fktz = "fktz:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
